manager: Make SELinux states a method
